Add basic packaging tasks to the dsl
This is a demo for what is needed for testing features that rely on 3rd
party utilities.

Another way to go about this is to use Puppet, unfortunately, I need
these added directly to Puppet's environment (ie, I need to use PE's
vendored `gem` utility). The location of additional PE packages also
depends on the install path of PE.

I hope this at least shows a use case, that needs more discussion.

Signed-off-by: Justin Stoller <justin@puppetlabs.com>